Added proper ftoa

+++ modules/include/utils.h
@@ -5,4 +5,4 @@
 
 
 double atof(const char* s);
-void ftoa(char* buff, float value);
+void ftoa(char* buff, double value);
